feat: rename files, replace watch with computed
