+++ src/components/SearchComponent.vue
